Intro and Syntax C# exercises

+++ IntroAndSyntax/04BeverageLabels/Program.cs
@@ -9,10 +9,10 @@ namespace _04BeverageLabels
         {
             string beverage = Console.ReadLine();
             int volume = int.Parse(Console.ReadLine());
-            int energyContent = int.Parse(Console.ReadLine());
-            int sugarContent = int.Parse(Console.ReadLine());
+            double energyContent = double.Parse(Console.ReadLine());
+			double sugarContent = double.Parse(Console.ReadLine());
 
-            double energy = (volume * energyContent)/100;
+            double energy = (volume * (double)energyContent)/100;
             double sugar = (volume * (double)sugarContent)/100;
 
             Console.WriteLine($"{volume}ml {beverage}:");
